port: Batch 4 - fix many high-frequency errors (Codec, VertexFormat, etc.)
NEW CLASSES (10):
- com.jcraft.jogg.SyncState (MC uses SyncState, not StreamState - my earlier mistake)
- com.sun.jna.Native (main JNA class)
- com.sun.jna.platform.win32.Kernel32Util
- io.netty.channel.epoll.EpollIoHandler
- io.netty.channel.kqueue.KQueueSocketChannel
- com.mojang.serialization.{DataResult, DynamicOps, MapCodec, Lifecycle}
- com.mojang.blaze3d.pipeline.{CompiledRenderPipeline, RenderPipeline}
- com.mojang.blaze3d.vertex.VertexFormatElement

MAJOR STUB UPDATES:
- Codec: full interface with 12 static fields (BOOL, BYTE, INT, LONG, FLOAT, DOUBLE, STRING, etc.)
  + ~200 default methods covering all common Codec operations
- VertexFormat: add Mode.indexCount(int), builder(), getVertexSize(), contains(), getElementName()
- Identifier: add CODEC and STREAM_CODEC static fields
- TextureFormat: add RED8, RED8I, and many more enum constants
- SignatureState: add INVALID constant
- MinecraftProfileTextures: add EMPTY field
- Window: getGuiScale() now returns int (MC 26.1.2 API change from double)
- vertex.PoseStack: add covariant last() returning vertex.PoseStack.Pose
  (Pose extends systems.PoseStack.Pose for Java covariant return type)

MINOR STUB UPDATES:
- Page.eos(), Page.granulepos()
- StreamState.pagein(Page)
- Block constructor (DspState)
- GpuBuffer.slice(long, long)
- CommandEncoder: clearColorAndDepthTextures, clearColorTexture, clearDepthTexture
- GpuDevice.precompilePipeline
- GpuQuery: add close() and getValue (returns OptionalLong-compatible)
- GpuTextureView.getWidth(int)
- ALCapabilities: add AL_EXT_source_distance_model field

+++ sources/teavm/java/com/mojang/blaze3d/platform/Window.java
@@ -45,7 +45,7 @@ public class Window {
         private int framerateLimit = 260;
 
         /** The GUI scale factor. */
-        private double guiScale = 1.0;
+        private int guiScale = 1;
 
         /** The framebuffer width (same as canvas drawing buffer width). */
         private int framebufferWidth;
@@ -290,11 +290,11 @@ public class Window {
 
         // ========== GUI Scale ==========
 
-        public double getGuiScale() {
+        public int getGuiScale() {
                 return guiScale;
         }
 
-        public void setGuiScale(double scale) {
+        public void setGuiScale(int scale) {
                 this.guiScale = scale;
         }
